Hadmard gradient
